Windows|64-bit: Build on MSVC 2015, 64-bit Qt 5.6; cleanup of external libs
EAX2 is not available in 64 bits, so the DirectSound plugin is
compiled without any EAX2 features when doing a 64 bit build.

LZSS is now compiled from source on all platforms.

Updated zlib to 1.2.8 on Windows, with prebuilt binaries included in
the repository (x86 and x64).

Fixed a couple of warnings about converting 64-bit values to 32-bit.

IssueID #1574

+++ doomsday/apps/client/include/def_main.h
@@ -40,7 +40,7 @@ struct Array : public std::vector<PODType>
         return !size();
     }
     int size() const {
-        return std::vector<PODType>::size();
+        return (int) std::vector<PODType>::size();
     }
     void clear() {
         _elements = nullptr;
